requirements.txt
anyio==3.6.2
APScheduler==3.6.3
autopep8==2.0.1
cachetools==4.2.2
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
flake8==6.0.0
Flask==2.2.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-telegram-bot==13.15
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tornado==6.1
tzdata==2022.7
tzlocal==4.2
Werkzeug==2.2.2
